Break down implementation into smaller modules
