+++ crates/cgp-component-macro-lib/src/getter_component/getter_field.rs
@@ -1,0 +1,9 @@
+use syn::token::Mut;
+use syn::{Ident, Type};
+
+pub struct GetterField {
+    pub field_name: Ident,
+    pub field_type: Type,
+    pub provider_type: Type,
+    pub field_mut: Option<Mut>,
+}
