+++ crates/cgp-component-macro-lib/src/getter_component/mod.rs
@@ -1,1 +1,7 @@
 pub mod derive;
+pub mod getter_field;
+pub mod parse;
+pub mod symbol;
+pub mod use_field;
+pub mod use_fields;
+pub mod with_provider;
