+++ crates/cgp-component-macro-lib/src/getter_component/symbol.rs
@@ -1,0 +1,9 @@
+use syn::{parse_quote, Type};
+
+pub fn symbol_from_string(value: &str) -> Type {
+    value
+        .chars()
+        .rfold(parse_quote! { Nil }, |tail, c: char| -> Type {
+            parse_quote!( Cons< Char< #c >, #tail > )
+        })
+}
